% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputSum.R
\name{OutputSum}
\alias{OutputSum}
\title{Outputting summarized information of a data set}
\usage{
OutputSum(input_data, ref_group, sum_col)
}
\arguments{
\item{input_data}{Data with planted_date of trees. This parameter should be a data frame.}

\item{ref_group}{The grouping basis of the data. This parameter should be a string.}

\item{sum_col}{The columns you want to generate summarized information.}
}
\value{
A data set with summarized information of specified columns.
The output will be a data frame.
}
\description{
This function generates a summary of specific columns.
}
\examples{
Specify a group basis and a column which you are interested in:
OutputSum(datateachr::vancouver_trees,"species_name","tree_age")

If you want to generate a summary with multiple columns:
OutputSum(datateachr::vancouver_trees,"neighbourhood_name",c("tree_age","diameter"))
}
