% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubGrpSel.R
\name{SubGrpSel}
\alias{SubGrpSel}
\title{Select subgroups in the data}
\usage{
SubGrpSel(input_data, col_name, chosen_grp)
}
\arguments{
\item{input_data}{Data with planted_date of trees. This parameter should be a data frame.}

\item{col_name}{The column which you will extract categories. This parameter should be a character.}

\item{chosen_grp}{The areas you are interested in. This parameter should be a character.}
}
\value{
A filtered data in the interested regions. The output will be a data frame.
}
\description{
This function pulls out the specified categories in the data.
}
\examples{
## If you want to choose data located in a specific area ("ARBUTUS-RIDGE") from
## the "vancouver_trees" dataset:

## {Syntax}
 \dontrun{
 SubGrpSel(vancouver_trees,"neighbourhood_name","ARBUTUS-RIDGE")
 }


##If you want to choose all "ACER" and "FRAXINUS" data from the "vancouver_trees" dataset:
## {Syntax}
 \dontrun{
 SubGrpSel(vancouver_trees, "genus_name", c("ACER","FRAXINUS"))
 }

}
