% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeAgeCal.R
\name{TreeAgeCal}
\alias{TreeAgeCal}
\title{Calculating the tree ages}
\usage{
TreeAgeCal(input_data, ref_date)
}
\arguments{
\item{input_data}{Data with planted_date of trees. This parameter should be a data frame.}

\item{ref_date}{The Reference date for calculating the tree age of each observation. This parameter should be a string.}
}
\value{
Input data with a column of the tree age of each observation.
The output will be a data frame.
}
\description{
This function calculates the tree age of each observation in the input.
}
\examples{
TreeAgeCal(datateachr::vancouver_trees,"2021-11-20")
TreeAgeCal(datateachr::vancouver_trees,"2014-05-12")

}
