% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputSum.R
\name{OutputSum}
\alias{OutputSum}
\title{Outputting a summary}
\usage{
OutputSum(input_data, ref_group, sum_col)
}
\arguments{
\item{input_data}{Data with planted_date of trees. This parameter should be a data frame.}

\item{ref_group}{The grouping basis of the data. This parameter should be a string.}

\item{sum_col}{The columns you want to generate summarized information.}
}
\value{
A data set with summarized information of specified columns.
The output will be a data frame.
}
\description{
This function generates a summary of specific columns from the input data.
Users can flip through the data easily.
}
\examples{
## If you want to group the data set by "species_name" and then output the summary about "diameter":
## {Syntax}
 \dontrun{
 OutputSum(vancouver_trees,"species_name","diameter")
 }


## If you want to generate a summary with multiple columns:
## {Syntax}
\dontrun{
OutputSum(vancouver_trees,"neighbourhood_name",c("diameter","height_range_id"))
}
}
