% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TmeDifCal.R
\name{TmeDifCal}
\alias{TmeDifCal}
\title{Calculating time length}
\usage{
TmeDifCal(input_data, start_date, end_date, n_col_name)
}
\arguments{
\item{input_data}{Data with planted_date of trees.
This parameter should be a data frame.}

\item{start_date}{One of columns in a data set.
This parameter should be a character, which must be one of the columns of input_data.}

\item{end_date}{A specific date for calculating time lengths.
This parameter should be a character.}

\item{n_col_name}{The name for the new column you create in this function.}
}
\value{
Input data with a column of time length of each observation.
The output will be a data frame.
}
\description{
This function calculates the tree age of each observation in the input.
}
\examples{
TmeDifCal(vancouver_trees,"date_planted","2021-11-20","tree_age")


}
